// document/interfaces/document-types.interfaces.ts
export enum DocumentType {
  IDENTITY = 'identity',
  PROOF_OF_ADDRESS = 'proof_of_address',
  POWER_OF_ATTORNEY = 'power_of_attorney',
  MEDICAL_PRESCRIPTION = 'medical_prescription',
  ANVISA_AUTHORIZATION = 'anvisa_authorization',
  OTHER = 'other',
}

export enum DocumentMimetype {
  JPEG = 'image/jpeg',
  PNG = 'image/png',
  PDF = 'application/pdf',
}